Update to clang 3.5

+++ building/compiler/t/t-time.c
@@ -6,8 +6,8 @@ int main(int argc, char *argv[])
 {
     time_t t = time(NULL);
     printf("time: %u\n", (int) t);
-    printf("sizeof(time_t): %d\n", sizeof(time_t));
-    printf("sizeof(off_t): %d\n", sizeof(off_t));
-    printf("sizeof(size_t): %d\n", sizeof(size_t));
+    printf("sizeof(time_t): %d\n", (int) sizeof(time_t));
+    printf("sizeof(off_t): %d\n", (int) sizeof(off_t));
+    printf("sizeof(size_t): %d\n", (int) sizeof(size_t));
     return 0;
 }
